+++ example/lib/main.dart
@@ -31,19 +31,45 @@ class HomePage extends StatefulWidget {
   State<HomePage> createState() => _HomePageState();
 }
 
-class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
-  final _bluetoothPrintPlus = BluetoothPrintPlus.instance;
+class _HomePageState extends State<HomePage> {
   BluetoothDevice? _device;
+  late StreamSubscription<bool> _isScanningSubscription;
+  late StreamSubscription<BlueState> _blueStateSubscription;
+  late StreamSubscription<ConnectState> _connectStateSubscription;
+  late StreamSubscription<Uint8List> _receivedDataSubscription;
+  late StreamSubscription<List<BluetoothDevice>> _scanResultsSubscription;
+  late List<BluetoothDevice> _scanResults;
 
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) => initBluetooth());
+    initBluetoothPrintPlusListen();
   }
 
-  Future<void> initBluetooth() async {
+  @override
+  void dispose() {
+    super.dispose();
+    _isScanningSubscription.cancel();
+    _blueStateSubscription.cancel();
+    _connectStateSubscription.cancel();
+    _receivedDataSubscription.cancel();
+    _scanResultsSubscription.cancel();
+    _scanResults.clear();
+    _device = null;
+  }
+
+  Future<void> initBluetoothPrintPlusListen() async {
+    /// listen scanResults
+    _scanResultsSubscription = BluetoothPrintPlus.scanResults.listen((event) {
+      if (mounted) {
+        setState(() {
+          _scanResults = event;
+        });
+      }
+    });
+
     /// listen isScanning
-    _bluetoothPrintPlus.isScanning.listen((event) {
+    _isScanningSubscription = BluetoothPrintPlus.isScanning.listen((event) {
       print('********** isScanning: $event **********');
       if (mounted) {
         setState(() {});
@@ -51,7 +77,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     });
 
     /// listen blue state
-    _bluetoothPrintPlus.blueState.listen((event) {
+    _blueStateSubscription = BluetoothPrintPlus.blueState.listen((event) {
       print('********** blueState change: $event **********');
       if (mounted) {
         setState(() {});
@@ -59,7 +85,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     });
 
     /// listen connect state
-    _bluetoothPrintPlus.connectState.listen((event) {
+    _connectStateSubscription = BluetoothPrintPlus.connectState.listen((event) {
       print('********** connectState change: $event **********');
       switch (event) {
         case ConnectState.connected:
@@ -80,7 +106,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     });
 
     /// listen received data
-    _bluetoothPrintPlus.receivedData.listen((data) {
+    _receivedDataSubscription = BluetoothPrintPlus.receivedData.listen((data) {
       print('********** received data: $data **********');
 
       /// do something...
@@ -90,78 +116,67 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('BluetoothPrintPlus'),
-      ),
-      body: SafeArea(
-        child: _bluetoothPrintPlus.isBlueOn
-            ? Column(
-                children: [
-                  Expanded(
-                      child: StreamBuilder<List<BluetoothDevice>>(
-                    stream: _bluetoothPrintPlus.scanResults,
-                    initialData: const [],
-                    builder: (c, snapshot) => ListView(
-                      children: snapshot.data!
-                          .map((device) => Container(
-                                padding: EdgeInsets.only(
-                                    left: 10, right: 10, bottom: 5),
-                                child: Row(
-                                  mainAxisAlignment:
-                                      MainAxisAlignment.spaceBetween,
-                                  children: [
-                                    Expanded(
-                                        child: Column(
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      children: [
-                                        Text(device.name),
-                                        Text(
-                                          device.address,
-                                          overflow: TextOverflow.ellipsis,
-                                          style: TextStyle(
-                                              fontSize: 12, color: Colors.grey),
-                                        ),
-                                        Divider(),
-                                      ],
-                                    )),
-                                    SizedBox(
-                                      width: 10,
-                                    ),
-                                    OutlinedButton(
-                                      onPressed: () async {
-                                        _device = device;
-                                        await _bluetoothPrintPlus
-                                            .connect(device);
-                                      },
-                                      child: const Text("connect"),
-                                    )
-                                  ],
-                                ),
-                              ))
-                          .toList(),
-                    ),
-                  )),
-                ],
-              )
-            : Center(
-                child: Text(
-                  "Bluetooth is turned off\nPlease turn on Bluetooth...",
-                  style: TextStyle(
-                      fontWeight: FontWeight.w700,
-                      fontSize: 16,
-                      color: Colors.red),
-                  textAlign: TextAlign.center,
-                ),
-              ),
-      ),
-      floatingActionButton:
-          _bluetoothPrintPlus.isBlueOn ? buildScanButton(context) : null,
-    );
+        appBar: AppBar(
+          title: const Text('BluetoothPrintPlus'),
+        ),
+        body: SafeArea(
+            child: BluetoothPrintPlus.isBlueOn
+                ? ListView(
+                    children: _scanResults
+                        .map((device) => Container(
+                              padding: EdgeInsets.only(
+                                  left: 10, right: 10, bottom: 5),
+                              child: Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
+                                children: [
+                                  Expanded(
+                                      child: Column(
+                                    crossAxisAlignment:
+                                        CrossAxisAlignment.start,
+                                    children: [
+                                      Text(device.name),
+                                      Text(
+                                        device.address,
+                                        overflow: TextOverflow.ellipsis,
+                                        style: TextStyle(
+                                            fontSize: 12, color: Colors.grey),
+                                      ),
+                                      Divider(),
+                                    ],
+                                  )),
+                                  SizedBox(
+                                    width: 10,
+                                  ),
+                                  OutlinedButton(
+                                    onPressed: () async {
+                                      _device = device;
+                                      await BluetoothPrintPlus.connect(device);
+                                    },
+                                    child: const Text("connect"),
+                                  )
+                                ],
+                              ),
+                            ))
+                        .toList(),
+                  )
+                : buildBlueOffWidget()),
+        floatingActionButton:
+            BluetoothPrintPlus.isBlueOn ? buildScanButton(context) : null);
+  }
+
+  Widget buildBlueOffWidget() {
+    return Center(
+        child: Text(
+      "Bluetooth is turned off\nPlease turn on Bluetooth...",
+      style: TextStyle(
+          fontWeight: FontWeight.w700, fontSize: 16, color: Colors.red),
+      textAlign: TextAlign.center,
+    ));
   }
 
   Widget buildScanButton(BuildContext context) {
-    if (_bluetoothPrintPlus.isScanningNow) {
+    if (BluetoothPrintPlus.isScanningNow) {
       return FloatingActionButton(
         onPressed: onStopPressed,
         backgroundColor: Colors.red,
@@ -177,7 +192,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
 
   Future onScanPressed() async {
     try {
-      await _bluetoothPrintPlus.startScan(timeout: Duration(seconds: 10));
+      await BluetoothPrintPlus.startScan(timeout: Duration(seconds: 10));
     } catch (e) {
       print("onScanPressed error: $e");
     }
@@ -185,7 +200,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
 
   Future onStopPressed() async {
     try {
-      _bluetoothPrintPlus.stopScan();
+      BluetoothPrintPlus.stopScan();
     } catch (e) {
       print("onStopPressed error: $e");
     }