+++ lib/src/blue.dart
@@ -5,39 +5,74 @@ import 'package:flutter/services.dart';
 import '../bluetooth_print_plus.dart';
 
 class BluetoothPrintPlus {
-  BluetoothPrintPlus._() {
-    _channel.setMethodCallHandler((MethodCall call) async {
-      _methodStreamController.add(call);
+  static bool _initialized = false;
+  static Future<dynamic> _initFlutterBluePlus() async {
+    if (_initialized) {
+      return;
+    }
+    _initialized = true;
+    // set platform method handler
+    _methodChannel.setMethodCallHandler((MethodCall call) async {
+      _methodStream.add(call);
     });
     _state.listen((event) {});
   }
-  static final BluetoothPrintPlus _instance = BluetoothPrintPlus._();
-  static BluetoothPrintPlus get instance => _instance;
-  static const MethodChannel _channel =
-      MethodChannel("bluetooth_print_plus/methods");
-  static const EventChannel _stateChannel =
-      EventChannel('bluetooth_print_plus/state');
-  final StreamController<MethodCall> _methodStreamController =
-      StreamController.broadcast();
-  static final _scanResults =
-      StreamControllerReEmit<List<BluetoothDevice>>(initialValue: []);
+
+  /// native platform methods channel
+  static final MethodChannel _methodChannel = const MethodChannel("bluetooth_print_plus/methods");
+
+  /// native platform events channel
+  static final EventChannel _stateChannel = const EventChannel('bluetooth_print_plus/state');
+  static final StreamController<MethodCall> _methodStream = StreamController.broadcast();
+
+  /// stream used for the scanResults public api
+  static final _scanResults = StreamControllerReEmit<List<BluetoothDevice>>(initialValue: []);
+
+  /// stream used for the isScanning public api
   static final _isScanning = StreamControllerReEmit<bool>(initialValue: false);
-  static final _connectState = StreamControllerReEmit<ConnectState>(
-      initialValue: ConnectState.disconnected);
-  static final _blueState =
-      StreamControllerReEmit<BlueState>(initialValue: BlueState.blueOn);
+
+  /// stream used for the isConnected public api
+  static final _connectState = StreamControllerReEmit<ConnectState>(initialValue: ConnectState.disconnected);
+
+  /// stream used for the isBlueOn public api
+  static final _blueState = StreamControllerReEmit<BlueState>(initialValue: BlueState.blueOn);
+
+  /// timeout for scanning that can be cancelled by stopScan
   static Timer? _scanTimeout;
 
-  Stream<MethodCall> get _methodStream => _methodStreamController.stream;
-  Stream<List<BluetoothDevice>> get scanResults => _scanResults.stream;
-  Stream<bool> get isScanning => _isScanning.stream;
-  Stream<ConnectState> get connectState => _connectState.stream;
-  Stream<BlueState> get blueState => _blueState.stream;
-  bool get isScanningNow => _isScanning.latestValue;
-  bool get isConnected => _connectState.latestValue == ConnectState.connected;
-  bool get isBlueOn => _blueState.latestValue == BlueState.blueOn;
+  /// a stream of scan results
+  /// - if you re-listen to the stream it re-emits the previous results
+  /// - the list contains all the results since the scan started
+  /// - the returned stream is never closed.
+  static Stream<List<BluetoothDevice>> get scanResults => _scanResults.stream;
+
+  /// returns whether we are scanning as a stream
+  static Stream<bool> get isScanning => _isScanning.stream;
+
+  /// returns connect state as a stream
+  static Stream<ConnectState> get connectState => _connectState.stream;
+
+  /// returns blue state as a stream
+  static Stream<BlueState> get blueState => _blueState.stream;
+
+  /// are we scanning right now?
+  static bool get isScanningNow => _isScanning.latestValue;
+
+  /// blue device is connected now?
+  static bool get isConnected => _connectState.latestValue == ConnectState.connected;
+
+  /// blue is on now?
+  static bool get isBlueOn => _blueState.latestValue == BlueState.blueOn;
+
+  /// the last known state
+  static int? _stateNow;
+
+  /// the subscription to the merged scan results stream
+  static StreamSubscription<BluetoothDevice?>? _scanSubscription;
 
   /// Start a scan for Bluetooth devices.
+  ///
+  /// [timeout] calls stopScan after a specified duration
   ///
   /// If a scan is already in progress, it is stopped first.
   ///
@@ -46,9 +81,10 @@ class BluetoothPrintPlus {
   /// The `timeout` parameter stops the scan after the specified duration.
   ///
   /// Returns the list of scanned devices.
-  Future startScan({
+  static Future startScan({
     Duration? timeout,
   }) async {
+    _initFlutterBluePlus();
     if (isScanningNow) {
       await stopScan();
     }
@@ -61,11 +97,11 @@ class BluetoothPrintPlus {
   /// If no scan is in progress, nothing happens.
   ///
   /// The `isScanning` stream is emitted with `false` when the scan is stopped.
-  Future stopScan() async {
+  static Future stopScan() async {
     if (isScanningNow) {
-      await _channel.invokeMethod('stopScan');
       _isScanning.add(false);
       _scanTimeout?.cancel();
+      await _methodChannel.invokeMethod('stopScan');
     } else {
       print("stopScan: already stopped");
     }
@@ -81,8 +117,8 @@ class BluetoothPrintPlus {
   ///
   /// The `connect` method returns immediately, and the connection status
   /// is reported on the `connectState` stream.
-  Future<dynamic> connect(BluetoothDevice device) async {
-    await _channel.invokeMethod('connect', device.toJson());
+  static Future<dynamic> connect(BluetoothDevice device) async {
+    await _methodChannel.invokeMethod('connect', device.toJson());
   }
 
   /// Disconnects from the currently connected Bluetooth device.
@@ -92,8 +128,8 @@ class BluetoothPrintPlus {
   /// The `connectState` stream is emitted with `disconnected` after a successful disconnection.
   ///
   /// Returns a `Future` that completes when the disconnection process is finished.
-  Future<dynamic> disconnect() async {
-    await _channel.invokeMethod('disconnect');
+  static Future<dynamic> disconnect() async {
+    await _methodChannel.invokeMethod('disconnect');
   }
 
   /// Sends data to the connected Bluetooth device.
@@ -104,13 +140,13 @@ class BluetoothPrintPlus {
   /// passing the data as an argument.
   ///
   /// Returns a `Future` that completes when the write operation is finished.
-  Future<dynamic> write(Uint8List? data) async {
-    await _channel.invokeMethod('write', {"data": data});
+  static Future<dynamic> write(Uint8List? data) async {
+    await _methodChannel.invokeMethod('write', {"data": data});
   }
 
   /// peripheral data feedback, receive and listen;
-  Stream<Uint8List> get receivedData async* {
-    yield* BluetoothPrintPlus.instance._methodStream
+  static Stream<Uint8List> get receivedData async* {
+    yield* BluetoothPrintPlus._methodStream.stream
         .where((m) => m.method == "ReceivedData")
         .map((m) {
       return m.arguments;
@@ -118,23 +154,14 @@ class BluetoothPrintPlus {
   }
 
   /// Gets the current state of the Bluetooth module
-  Stream<int> get _state async* {
-    yield await _channel.invokeMethod('state').then((s) {
-      if (s <= 1) {
-        if (s == 0) {
-          _blueState.add(BlueState.blueOn);
-        } else if (s == 1) {
-          _blueState.add(BlueState.blueOff);
-        }
-      } else {
-        if (s == 2) {
-          _connectState.add(ConnectState.connected);
-        } else if (s == 3) {
-          _connectState.add(ConnectState.disconnected);
-        }
+  static Stream<int> get _state async* {
+    if (_stateNow == null) {
+      var result = await _methodChannel.invokeMethod('state');
+      // update _adapterStateNow if it is still null after the await
+      if (_stateNow == null) {
+        _stateNow = result;
       }
-      return 1;
-    });
+    }
 
     yield* _stateChannel.receiveBroadcastStream().map((s) {
       if (s <= 1) {
@@ -168,20 +195,20 @@ class BluetoothPrintPlus {
   ///
   /// If the scan fails (for example, if the device does not have Bluetooth
   /// capabilities), the stream will emit an error and then close.
-  Stream<BluetoothDevice> _scan({Duration? timeout}) async* {
+  static Stream<BluetoothDevice> _scan({Duration? timeout}) async* {
     // Emit to isScanning
     _isScanning.add(true);
     // Clear scan results list
     _scanResults.add(<BluetoothDevice>[]);
     // invoke startScan method
-    await _channel
+    await _methodChannel
         .invokeMethod('startScan')
         .onError((error, stackTrace) => _isScanning.add(false));
     if (timeout != null) {
       _scanTimeout = Timer(timeout, stopScan);
     }
 
-    yield* BluetoothPrintPlus.instance._methodStream
+    yield* BluetoothPrintPlus._methodStream.stream
         .where((m) => m.method == "ScanResult")
         .map((m) => m.arguments)
         .map((map) {