+++ lib/src/utils.dart
@@ -1,5 +1,9 @@
 import 'dart:async';
 
+// This is a reimplementation of BehaviorSubject from RxDart library.
+// It is essentially a stream but:
+//  1. we cache the latestValue of the stream
+//  2. the "latestValue" is re-emitted whenever the stream is listened to
 class StreamControllerReEmit<T> {
   T latestValue;