+++ CHANGELOG.md
@@ -1,21 +1,51 @@
-## 【2.4.0】- Changed
-- 优化安卓连接 - improve Android Bluetooth `connect`;
-- 区分蓝牙状态和连接状态 - `Distinguish` between Bluetooth status and connection status;
-- 添加一些属性，eg：isBlueOn、isConnected - `Add` some properties, eg：isBlueOn、isConnected;
-- 优化 example - improve `example`;
+## V2.4.1
+##### 修改点
+- 优化流监听取消; 
+- 优化 BluetoothPrintPlus 类; 
+- 添加代码注释;
+- 优化 example; 
 
-## 【2.3.4】- Changed
-- 优化安卓蓝牙扫描 - improve Android Bluetooth `scan`;
-- 优化 example - improve `example`;
+##### Changed
+- improve `stream` listen and cancel;
+- improve `BluetoothPrintPlus` class;
+- Add code `comments`;
+- improve `example`;
 
-## 【2.3.3】- Changed
-- 优化状态改变监听 - improve `status` change listen;
-- 添加蓝牙开启关闭监听 - Add Bluetooth, `turn on/off` listen;
-- 添加外设数据反馈接收监听 - Add peripheral `data feedback`, receive and listen;
-- 调整一些代码 - Adjust some code;
+##  V2.4.0
+##### 修改点
+- 优化安卓连接; 
+- 区分蓝牙状态和连接状态; 
+- 添加一些属性，eg：isBlueOn、isConnected; 
+- 优化 example; 
+##### Changed
+- improve Android Bluetooth `connect`;
+- `Distinguish` between Bluetooth status and connection status;
+- `Add` some properties, eg：isBlueOn、isConnected;
+- improve `example`;
+
+## V2.3.4
+##### 修改点
+- 优化安卓蓝牙扫描; 
+- 优化 example; 
+
+##### Changed
+- improve Android Bluetooth `scan`;
+- improve `example`;
+
+## V2.3.3
+##### 修改点
+- 优化状态改变监听; 
+- 添加蓝牙开启关闭监听; 
+- 添加外设数据反馈接收监听; 
+- 调整一些代码; 
+##### Changed
+- improve `status` change listen;
+- Add Bluetooth, `turn on/off` listen;
+- Add peripheral `data feedback`, receive and listen;
+- Adjust some code;
 
 ...
 
-## 0.0.1
+## V0.0.1
 
 - TODO: Describe initial release.